+++ Functions/__pycache__/token_manipulation.cpython-36.pyc
