+++ Functions/adapter.py
