+++ Functions/global_variables.py
