+++ Functions/structural_calculation.py
