+++ Functions/structure_generation.py
