+++ Functions/token_manipulation.py
