+++ Functions/vector_calculation.py
