+++ Version Details 1.7.1.txt
