+++ templates/new-output.html
@@ -1,0 +1,69 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<title>Semantic Tree - Concept Map</title>
+
+	<script src="{{url_for('static', filename='js/jquery-2.1.1.min.js')}}"></script>
+	<script type="text/javascript" src="{{url_for('static', filename='js/development.js')}}"></script>
+	<script src="{{url_for('static', filename='js/cytoscape.js')}}"></script>
+    <script src="{{url_for('static', filename='layouts/cose-bilkent/cytoscape-cose-bilkent.js')}}"></script>
+	<script src="{{url_for('static', filename='js/cytoscape-cxtmenu.js')}}"></script>
+	<script type="text/javascript" src="{{url_for('static', filename='js/variable_list.js')}}"></script>
+	<script type="text/javascript" src="{{url_for('static', filename='js/concept_map_rendering_functions.js')}}"></script>
+
+
+	<link href="https://fonts.googleapis.com/css?family=Fjalla+One" rel="stylesheet">
+	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+	<link rel="stylesheet" type="text/css" href="{{url_for('static', filename='styles/new-output.css')}}">
+	<link rel="shortcut icon" href="/">
+
+
+</head>
+<body>
+	<div class="grid-container">
+		<div id="viewport"></div>
+		<div id="config-area"></div>
+		<div id="profile-area"></div>
+		<div id='profile-icon'><i class="fa fa-sitemap"></i></div>
+		<div id="profile-title">Semantic Tree</div>
+
+		<!-- <div id="button-area"></div> -->
+		<div class="btn" id='button-reset'><button onclick="reset()"><i class="fa fa-refresh"></i> Reset</button></div>
+		<div class="btn" id="button-og-text"><button onclick="showText(0,orignalText.length)"><i class="fa fa-file-text-o"></i> Original Text</button></div>
+		<div id="origin-text-area">
+			<div id="origin-text"></div>
+		</div>
+
+		<!-- Control Area -->
+		<div id="slider-area">
+			<div>
+				<p>Threshhold</p>
+				<input type="range" min="0", max="1", step="0.01" value="0.75", id="threshhold">
+				<span id="threshhold-output"></span>
+			</div>
+			<div>
+				<p>Maximum Connections</p>
+				<input type="range" min="2", max="50", value="5", id="maxConnections">
+				<span id="maxConnections-output"></span>
+			</div>
+			<div class="btn">
+				<button onclick="changeParameters()" class="btn">Apply parameters</button>
+			</div>
+		</div>
+
+		<!-- <div id="back-area"></div> -->
+		<div class="btn" id="back-button-back"><button onclick="location.href='/input_page'"><i class="fa fa-long-arrow-left"></i> Back</button></div>
+
+		<div id="map-area"></div>
+		<div id="map-zoom" class="btn-group">
+			<button class='zoom' onclick='____'>+</button>
+			<button class='zoom' onclick='____'>-</button>
+		</div>
+	</div>
+	<script type="text/javascript">
+		var cy=create_core_obj(true);
+		LoadFile();
+		setLayout("cose-bilkent", cy);
+	</script>
+</body>
+</html>
